test: 무비 데이, 할인 시간 테스트 코드 작성
